Assignment 4/requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
apache-airflow-providers-apache-hdfs
SQLAlchemy
pyspark
Flask
WTForms
argparse
pandas